# docs/swagger.yaml
definitions:
  disgm.Guild:
    properties:
      afk_channel_id:
        description: Optional ID of AFK channel
        type: string
      afk_timeout:
        description: AFK timeout in seconds
        type: integer
      application_id:
        description: Optional application ID if bot-created
        type: string
      approximate_member_count:
        description: Optional approximate number of members
        type: integer
      approximate_presence_count:
        description: Optional approximate non-offline members
        type: integer
      banner:
        description: Optional banner hash
        type: string
      default_message_notifications:
        description: Default message notifications level
        type: integer
      description:
        description: Optional description of the guild
        type: string
      discovery_splash:
        description: Optional discovery splash hash for discoverable guilds
        type: string
      emojis:
        description: Custom guild emojis
        items:
          $ref: '#/definitions/models.Emoji'
        type: array
      explicit_content_filter:
        description: Explicit content filter level
        type: integer
      features:
        description: Enabled guild features
        items:
          type: string
        type: array
      icon:
        description: Optional icon hash
        type: string
      icon_hash:
        description: Optional icon hash returned in the template object
        type: string
      id:
        description: Snowflake ID of the guild
        type: string
      max_members:
        description: Maximum number of members for the guild
        type: integer
      max_presences:
        description: Optional maximum presences for the guild
        type: integer
      max_stage_video_channel_users:
        description: Maximum users in a stage video channel
        type: integer
      max_video_channel_users:
        description: Maximum users in a video channel
        type: integer
      mfa_level:
        description: Required MFA level for the guild
        type: integer
      name:
        description: Name of the guild (2-100 characters)
        type: string
      nsfw_level:
        description: NSFW level of the guild
        type: integer
      owner:
        description: Optional flag indicating if the user is the owner of the guild
        type: boolean
      owner_id:
        description: ID of the owner
        type: string
      permissions:
        description: Optional total permissions for the user in the guild
        type: string
      preferred_locale:
        description: Preferred locale of the community guild
        type: string
      premium_progress_bar_enabled:
        description: Flag for boost progress bar enabled
        type: boolean
      premium_subscription_count:
        description: Optional number of boosts
        type: integer
      premium_tier:
        description: Premium tier (Server Boost level)
        type: integer
      public_updates_channel_id:
        description: Optional public updates channel ID
        type: string
      region:
        description: Optional voice region ID for the guild (deprecated)
        type: string
      roles:
        description: Roles in the guild
        items:
          $ref: '#/definitions/models.Role'
        type: array
      rules_channel_id:
        description: Optional channel ID for community rules
        type: string
      safety_alerts_channel_id:
        description: Optional channel ID for safety alerts
        type: string
      splash:
        description: Optional splash hash
        type: string
      stickers:
        description: Custom guild stickers
        items:
          $ref: '#/definitions/models.Sticker'
        type: array
      system_channel_flags:
        description: System channel flags
        type: integer
      system_channel_id:
        description: Optional system channel ID for notices
        type: string
      vanity_url_code:
        description: Optional vanity URL code for the guild
        type: string
      verification_level:
        description: Verification level required for the guild
        type: integer
      welcome_screen:
        allOf:
        - $ref: '#/definitions/models.WelcomeScreen'
        description: Optional welcome screen object
      widget_channel_id:
        description: Optional channel ID for widget invite
        type: string
      widget_enabled:
        description: Optional flag indicating if the server widget is enabled
        type: boolean
    type: object
...
  models.AvatarDecorationData:
    properties:
      decoration:
        description: Example field for decoration
        type: string
    type: object
...
  models.Emoji:
    properties:
      animated:
        description: Whether the emoji is animated
        type: boolean
      id:
        description: Define fields for Emoji structure based on your needs
        type: string
      managed:
        description: Whether the emoji is managed by an application
        type: boolean
      name:
        description: Name of the emoji
        type: string
      requires_colons:
        description: Whether the emoji requires colons
        type: boolean
      roles:
        description: Roles allowed to use the emoji
        items:
          $ref: '#/definitions/models.Role'
        type: array
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Optional user object that created the emoji
    type: object
  models.GuildBan:
    properties:
      reason:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.Role:
    properties:
      color:
        description: Color of the role
        type: integer
      hoist:
        description: Whether the role is hoisted in the user list
        type: boolean
      id:
        description: Define fields for Role structure based on your needs
        type: string
      managed:
        description: Whether the role is managed by an application
        type: boolean
      mentionable:
        description: Whether the role is mentionable
        type: boolean
      name:
        description: Name of the role
        type: string
      permissions:
        description: Permissions for the role
        type: string
      position:
        description: Position of the role
        type: integer
    type: object
...
  models.Sticker:
    properties:
      description:
        description: Optional description of the sticker
        type: string
      format_type:
        description: Format type of the sticker
        type: integer
      id:
        description: Define fields for Sticker structure based on your needs
        type: string
      name:
        description: Name of the sticker
        type: string
      pack_id:
        description: ID of the sticker pack
        type: string
    type: object
...
  models.User:
    properties:
      accent_color:
        description: Optional banner color as an integer
        type: integer
      avatar:
        description: Optional avatar hash of the user
        type: string
      avatar_decoration_data:
        allOf:
        - $ref: '#/definitions/models.AvatarDecorationData'
        description: Optional avatar decoration data
      banner:
        description: Optional banner hash of the user
        type: string
      bot:
        description: Optional flag indicating if the user is a bot
        type: boolean
      discriminator:
        description: User's Discord tag (four-digit identifier)
        type: string
      email:
        description: Optional user's email
        type: string
      flags:
        description: Optional flags on the user's account
        type: integer
      global_name:
        description: Optional display name (for bots, the application name)
        type: string
      id:
        description: Snowflake ID of the user
        type: string
      locale:
        description: Optional user's chosen language
        type: string
      mfa_enabled:
        description: Optional flag indicating if 2FA is enabled
        type: boolean
      premium_type:
        description: Optional Nitro subscription type
        type: integer
      public_flags:
        description: Optional public flags on the user's account
        type: integer
      system:
        description: Optional flag indicating if the user is a system user
        type: boolean
      username:
        description: Username of the user (not unique)
        type: string
      verified:
        description: Optional flag indicating if the email is verified
        type: boolean
    type: object
  models.WelcomeChannel:
    properties:
      channel_id:
        description: ID of the channel
        type: string
      description:
        description: Optional description for the channel
        type: string
      emoji:
        description: Optional emoji for the channel
        type: string
    type: object
  models.WelcomeScreen:
    properties:
      description:
        description: Define fields relevant to the welcome screen as needed
        type: string
      welcome_channels:
        description: Channels to show on the welcome screen
        items:
          $ref: '#/definitions/models.WelcomeChannel'
        type: array
    type: object
host: localhost:90
info:
  contact: {}
  description: API for managing Discord guilds using DiscordGo and Fiber.
  title: Discord Guild Management API
  version: "1.0"
paths:
  /api/guild:
    get:
      description: Retrieve the guild information.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/disgm.Guild'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild
      tags:
      - Guild
  /api/guild/bans:
    get:
      description: Retrieve all banned users from the guild.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GuildBan'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Bans
      tags:
      - Bans
  /api/guild/bans/{userid}:
    delete:
      description: Remove a ban for a user in the guild.
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Remove Guild Ban
      tags:
      - Bans
    get:
      description: Retrieve a specific banned user by user ID.
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GuildBan'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get Guild Ban
      tags:
      - Bans
    put:
      description: Ban a user from the guild.
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Add Guild Ban
      tags:
      - Bans
  /api/guild/bulk-ban:
    post:
      description: Ban multiple users in the guild at once.
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Bulk Ban Members
      tags:
      - Bans
...
  /api/guild/interactions/{interactionid}/{interactiontoken}/callback:
    post:
      description: Handle interaction callback for a specific interaction.
      parameters:
      - description: Interaction ID
        in: path
        name: interactionid
        required: true
        type: string
      - description: Interaction Token
        in: path
        name: interactiontoken
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create Interaction Callback
      tags:
      - Interactions
...
swagger: "2.0"